Combine the states and paths into landscape
