+++ neet/__init__.py
@@ -2,5 +2,5 @@
 # Use of this source code is governed by a MIT
 # license that can be found in the LICENSE file.
 from .interfaces import *
-from .states import *
-from .paths import *
+from .landscape import *
+
